pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "solace-agent-mesh"
dynamic = ["version"]
authors = [
  { name="Edward Funnekotter", email="edward.funnekotter@solace.com" },
  { name="Greg Meldrum", email="greg.meldrum@solace.com" },
  { name="Cyrus Mobini", email="cyrus.mobini@solace.com" },
]
description = "Solace Agent Mesh is an EDA AI-first platform powered by Solace"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "boto3~=1.34.122",
    "langchain-core~=0.3.0",
    "langchain~=0.3.0",
    "PyYAML~=6.0.1",
    "Requests~=2.32.3",
    "solace_pubsubplus~=1.9.0",
    "click~=8.1.7",
    "plotly~=5.24.1",
    "html2text~=2024.2.26",
    "beautifulsoup4~=4.12.3",
    "azure-identity~=1.17.1",
    "msgraph-core~=1.1.2",
    "msgraph-sdk~=1.5.4",
    "microsoft-kiota-abstractions==1.9.0",
    "microsoft-kiota-authentication-azure==1.9.0",
    "microsoft-kiota-http==1.9.0",
    "microsoft-kiota-serialization-form==1.9.0",
    "microsoft-kiota-serialization-json==1.9.0",
    "microsoft-kiota-serialization-multipart==1.9.0",
    "microsoft-kiota-serialization-text==1.9.0",
    "Flask~=3.0.3",
    "jq~=1.8.0",
    "litellm~=1.51.3",
    "ruamel.yaml~=0.18.6",
    "Flask-SocketIO~=5.4.1",
    "kaleido~=0.2.1",
    "mysql_connector_repackaged==0.3.1",
    "python_dateutil==2.9.0.post0",
    "PyYAML~=6.0.1",
    "pytest~=8.3.1",
    "pytest-cov~=5.0.0",
    "build~=1.2.2.post1",
    "duckduckgo_search~=6.3.7",
    "markitdown~=0.0.1a3",
    "pyperclip~=1.9.0",
    "solace-ai-connector @ git+https://github.com/SolaceDev/solace-ai-connector.git@main",
    "solace-ai-connector-slack @ git+https://github.com/SolaceLabs/solace-ai-connector-slack.git@main",
    "solace-ai-connector-rest @ git+https://github.com/SolaceLabs/solace-ai-connector-rest.git@main",
    "solace-ai-connector-web @ git+https://github.com/SolaceLabs/solace-ai-connector-web.git@main",
]

[project.urls]
homepage = "https://github.com/SolaceLabs/solace-agent-mesh"
repository = "https://github.com/SolaceLabs/solace-agent-mesh"
documentation = "https://github.com/SolaceLabs/solace-agent-mesh/blob/main/docs/docs/index.md"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel.force-include]
"src" = "solace_agent_mesh/"
"configs" = "solace_agent_mesh/configs"
"templates" = "solace_agent_mesh/templates"
"cli" = "solace_agent_mesh/cli"

[tool.hatch.build.targets.sdist.force-include]
"web-visualizer/dist" = "solace_agent_mesh/assets/web-visualizer"

[project.scripts]
solace-agent-mesh = "solace_agent_mesh.cli.main:main"
sam = "solace_agent_mesh.cli.main:main"

[tool.hatch.build.targets.wheel]
packages = ["solace_agent_mesh"]


[tool.hatch.version]
path = "cli/__init__.py"

[tool.ruff]
lint.select = ["E4", "E7", "E9", "F"]
lint.ignore = ["F401", "E731"]

[tool.hatch.envs.hatch-test]
installer = "pip"


[[tool.hatch.envs.hatch-test.matrix]]
python = ["3.11", "3.12"]